refactor: rename trinosql to trino
