+++ test/parser/trino/contextCollect/entityCollector.test.ts
@@ -3,7 +3,7 @@ import path from 'path';
 import TrinoSQL from 'src/parser/trino';
 import { TrinoSqlSplitListener, TrinoEntityCollector } from 'src/parser/trino';
 import { ParseTreeListener } from 'antlr4ng';
-import { TrinoSqlListener } from 'src/lib/trinosql/TrinoSqlListener';
+import { TrinoSqlListener } from 'src/lib/trino/TrinoSqlListener';
 import { EntityContextType } from 'src/parser/common/basic-parser-types';
 import { StmtContextType } from 'src/parser/common/entityCollector';
